% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cividis-scales.R
\name{cividis_pal}
\alias{cividis_pal}
\title{cividis palette (discrete)}
\usage{
cividis_pal(alpha = 1, begin = 0, end = 1, direction = 1,
  option = "D")
}
\arguments{
\item{alpha}{pass through parameter to \code{cividis}}

\item{begin}{The (corrected) hue in [0,1] at which the cividis colormap begins.}

\item{end}{The (corrected) hue in [0,1] at which the cividis colormap ends.}

\item{direction}{Sets the order of colors in the scale. If 1, the default, colors
are ordered from darkest to lightest. If -1, the order of colors is reversed.}

\item{option}{A character string indicating the colormap option to use. Four
options are available: "magma" (or "A"), "inferno" (or "B"), "plasma" (or "C"),
and "cividis" (or "D", the default option).}
}
\description{
cividis palette (discrete)
}
\details{
See \link[cividis]{cividis} for more information on the color scale.
}
\examples{
library(scales)
show_col(cividis_pal()(10))


}
