% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cividis-scales.R
\name{scale_color_cividis}
\alias{scale_color_cividis}
\alias{scale_colour_cividis}
\alias{scale_color_cividis}
\alias{scale_fill_cividis}
\title{cividis color scales}
\usage{
scale_color_cividis(..., alpha = 1, begin = 0, end = 1, direction = 1,
  discrete = FALSE, option = "V")

scale_colour_cividis(..., alpha = 1, begin = 0, end = 1, direction = 1,
  discrete = FALSE, option = "V")

scale_fill_cividis(..., alpha = 1, begin = 0, end = 1, direction = 1,
  discrete = FALSE, option = "V")
}
\arguments{
\item{...}{parameters to \code{discrete_scale} or \code{scale_fill_gradientn}}

\item{alpha}{pass through parameter to \code{cividis}}

\item{begin}{The (corrected) hue in [0,1] at which the cividis colormap begins.}

\item{end}{The (corrected) hue in [0,1] at which the cividis colormap ends.}

\item{direction}{Sets the order of colors in the scale. If 1, the default, colors
are as output by \link[cividis]{cividis_pal}. If -1, the order of colors is reversed.}

\item{discrete}{generate a discrete palette? (default: \code{FALSE} - generate continuous palette)}

\item{option}{A character string indicating the colormap option to use. At the moment, only option "cividis" (or "V") is available.}
}
\description{
Uses the cividis color scale.
}
\details{
For \code{discrete == FALSE} (the default) all other arguments are as to
\link[ggplot2]{scale_fill_gradientn} or \link[ggplot2]{scale_color_gradientn}.
Otherwise the function will return a \code{discrete_scale} with the plot-computed
number of colors.

See \code{cividis} for more information on the color scale.
}
\examples{
library(ggplot2)

# ripped from the pages of ggplot2
p <- ggplot(mtcars, aes(wt, mpg))
p + geom_point(size=4, aes(colour = factor(cyl))) +
    scale_color_cividis(discrete=TRUE) +
    theme_bw()

# ripped from the pages of ggplot2
dsub <- subset(diamonds, x > 5 & x < 6 & y > 5 & y < 6)
dsub$diff <- with(dsub, sqrt(abs(x-y))* sign(x-y))
d <- ggplot(dsub, aes(x, y, colour=diff)) + geom_point()
d + scale_color_cividis() + theme_bw()


# ripped from the main viridis example
dat <- data.frame(x = rnorm(10000), y = rnorm(10000))

ggplot(dat, aes(x = x, y = y)) +
  geom_hex() + coord_fixed() +
  scale_fill_cividis() + theme_bw()

library(ggplot2)
library(MASS)
library(gridExtra)

data("geyser", package="MASS")

ggplot(geyser, aes(x = duration, y = waiting)) +
  xlim(0.5, 6) + ylim(40, 110) +
  stat_density2d(aes(fill = ..level..), geom="polygon") +
  theme_bw() +
  theme(panel.grid=element_blank()) -> gg

grid.arrange(
  gg + scale_fill_cividis(option="V") + labs(x="Virdis V", y=NULL),
  ncol=2, nrow=2
)

}
\author{
Marco Sciaini \email{msciain@uni-goettingen.de} / \href{https://twitter.com/msciain}{@msciain},
        C?dric Scherer \email{scherer@izw-berlin.de} / \href{https://twitter.com/CedScherer}{@CedScherer}
}
