% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cividis-colormap.R
\docType{data}
\name{cividis.map}
\alias{cividis.map}
\title{Original 'cividis' color map}
\format{A data frame with 1024 rows and 4 variables:
\itemize{
  \item R: Red value
  \item G: Green value
  \item B: Blue value
}}
\usage{
cividis.map
}
\description{
A dataset containing the original RGB values of the default Matplotlib color
 map ('cividis'). Source: \url{https://github.com/pnnl/cmaputil/blob/master/colormaps/cividis.txt}.
}
\keyword{datasets}
